# 2017-05-31-Docker_Refresher-Zero_to_Prod/demo-webapp/swagger/swagger.yaml
consumes:
- application/json
- application/xml
- application/gob
- application/x-gob
definitions:
  WsnPost:
    description: A post (default view)
    example:
      body: Non suscipit.
      id: 1.4570343e+18
      title: Quo et est nihil porro.
      userId: 5.5963625e+18
    properties:
      body:
        description: post body
        example: Non suscipit.
        type: string
      id:
        description: Unique post ID
        example: 1.4570343e+18
        format: int64
        type: integer
      title:
        description: post title
        example: Quo et est nihil porro.
        type: string
      userId:
        description: Owner's ID
        example: 5.5963625e+18
        format: int64
        type: integer
    required:
    - id
    - title
    - body
    - userId
    title: 'Mediatype identifier: application/vnd.wsn.post+json; view=default'
    type: object
...
host: localhost:8000
info:
  description: A terrible social network
  title: The worst social network
  version: ""
paths:
...
  /posts/{postID}:
    get:
      description: Get post by id
      operationId: post#show
      parameters:
      - description: Post ID
        in: path
        name: postID
        required: true
        type: integer
      produces:
      - application/vnd.wsn.post+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/WsnPost'
        "404":
          description: Not Found
      schemes:
      - http
      summary: show post
      tags:
      - post
...
produces:
- application/json
- application/xml
- application/gob
- application/x-gob
...
schemes:
- http
swagger: "2.0"